+++ volume.h
@@ -1,10 +1,12 @@
 #ifndef VOLUME_H
 #define VOLUME_H
 
+#include <stdint.h>
+
 enum VolumeIcon { IC_SPEAKER, IC_HEADPHONE, IC_BT_HEADSET, IC_MUTE };
 
-short get_volume(void);
-short get_mute(void);
-int get_volume_icon_type(void);
+uint8_t get_volume(void);
+uint8_t get_mute(void);
+uint8_t get_volume_icon_type(void);
 
 #endif // VOLUME_H
